server/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.4.2
python-multipart>=0.0.6
uuid>=1.30
playwright>=1.40.0
openai>=1.3.0
typing-extensions>=4.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
browserbase>=1.0.0 